# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/recipes/insert":{"post":{"summary":"Insert Recipe","operationId":"insert_recipe_recipes_insert_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recipes/list/all":{"get":{"summary":"Get All Recipes","operationId":"get_all_recipes_recipes_list_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/users/function/add":{"get":{"summary":"Populate User Accounts","operationId":"populate_user_accounts_users_function_add_get","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"},{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"query"},{"required":true,"schema":{"title":"Password","type":"string"},"name":"password","in":"query"},{"required":true,"schema":{"$ref":"#/components/schemas/UserType"},"name":"type","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/datamodel/add":{"post":{"summary":"Populate Login Without Service","operationId":"populate_login_without_service_users_datamodel_add_post","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"},{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"query"},{"required":true,"schema":{"title":"Password","type":"string"},"name":"password","in":"query"},{"required":true,"schema":{"$ref":"#/components/schemas/UserType"},"name":"type","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/add/profile":{"post":{"summary":"Add Profile Login","operationId":"add_profile_login_users_add_profile_post","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"},{"required":true,"schema":{"title":"Firstname","type":"string"},"name":"firstname","in":"query"},{"required":true,"schema":{"title":"Lastname","type":"string"},"name":"lastname","in":"query"},{"required":true,"schema":{"title":"Middle","type":"string"},"name":"middle","in":"query"},{"required":true,"schema":{"title":"Bday","type":"string","format":"date"},"name":"bday","in":"query"},{"required":true,"schema":{"title":"Pos","type":"string"},"name":"pos","in":"query"},{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"query"},{"required":true,"schema":{"title":"Password","type":"string"},"name":"password","in":"query"},{"required":true,"schema":{"$ref":"#/components/schemas/UserType"},"name":"type","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/add/model/profile":{"post":{"summary":"Add Profile Login Models","operationId":"add_profile_login_models_users_add_model_profile_post","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"},{"required":true,"schema":{"title":"Date Created","type":"string","format":"date"},"name":"date_created","in":"query"},{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"query"},{"required":true,"schema":{"title":"Password","type":"string"},"name":"password","in":"query"},{"required":true,"schema":{"$ref":"#/components/schemas/UserType"},"name":"type","in":"query"},{"required":true,"schema":{"title":"Firstname","type":"string"},"name":"firstname","in":"query"},{"required":true,"schema":{"title":"Lastname","type":"string"},"name":"lastname","in":"query"},{"required":true,"schema":{"title":"Middle","type":"string"},"name":"middle","in":"query"},{"required":true,"schema":{"title":"Bday","type":"string","format":"date"},"name":"bday","in":"query"},{"required":true,"schema":{"title":"Pos","type":"string"},"name":"pos","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/insert":{"post":{"summary":"Insert Post Feedback","operationId":"insert_post_feedback_posts_insert_post","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"},{"required":true,"schema":{"title":"Feedback","type":"string"},"name":"feedback","in":"query"},{"required":true,"schema":{"$ref":"#/components/schemas/RecipeRating"},"name":"rating","in":"query"},{"required":true,"schema":{"title":"Userid","type":"string","format":"uuid"},"name":"userId","in":"query"},{"required":true,"schema":{"title":"Date Posted","type":"string","format":"date"},"name":"date_posted","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login/query":{"get":{"summary":"Login With Query","operationId":"login_with_query_login_query_get","parameters":[{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"query"},{"required":true,"schema":{"title":"Password","type":"string"},"name":"password","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login/model":{"post":{"summary":"Login With Model","operationId":"login_with_model_login_model_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/login/details/list":{"get":{"summary":"List Login Details","operationId":"list_login_details_admin_login_details_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/user/profiles/list":{"get":{"summary":"List User Profiles","operationId":"list_user_profiles_admin_user_profiles_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/keyword/insert":{"post":{"summary":"Insert Recipe Keywords","operationId":"insert_recipe_keywords_keyword_insert_post","parameters":[{"required":true,"schema":{"title":"Rid","type":"string","format":"uuid"},"name":"rid","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Keywords","type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/keyword/add":{"post":{"summary":"Add Recipe Keyword","operationId":"add_recipe_keyword_keyword_add_post","parameters":[{"required":true,"schema":{"title":"Rid","type":"string","format":"uuid"},"name":"rid","in":"query"},{"required":true,"schema":{"title":"Keyword","type":"string"},"name":"keyword","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/keyword/get":{"post":{"summary":"Get Recipe Keywords","operationId":"get_recipe_keywords_keyword_get_post","parameters":[{"required":true,"schema":{"title":"Rid","type":"string","format":"uuid"},"name":"rid","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/keyword/list":{"get":{"summary":"Get All Keywords","operationId":"get_all_keywords_keyword_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/logs/visitors/list":{"get":{"summary":"List Logs Visitors","operationId":"list_logs_visitors_admin_logs_visitors_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/complaint/recipe":{"post":{"summary":"Report Recipe","operationId":"report_recipe_complaint_recipe_post","parameters":[{"required":true,"schema":{"title":"Rid","type":"string","format":"uuid"},"name":"rid","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/complaint/list/all":{"get":{"summary":"List Defective Recipes","operationId":"list_defective_recipes_complaint_list_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Category":{"title":"Category","enum":["breakfast","lunch","dinner","appetizer","salad","entree","side_dish","pastry","dessert","snack","soup","holiday","vegetarian","cookbook"],"type":"string","description":"An enumeration."},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"IngredientReq":{"title":"IngredientReq","required":["id","name","qty","measure"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"},"qty":{"title":"Qty","type":"integer"},"measure":{"title":"Measure","type":"string"}}},"LoginReq":{"title":"LoginReq","required":["username","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"}}},"Origin":{"title":"Origin","enum":["asian","mediterranean","mid_eastern","african","pacific","south_american","european","jewish","carribean"],"type":"string","description":"An enumeration."},"RecipeRating":{"title":"RecipeRating","enum":["1","2","3","4","5"],"type":"string","description":"An enumeration."},"RecipeReq":{"title":"RecipeReq","required":["id","name","ingredients","cat","orig"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"},"ingredients":{"title":"Ingredients","type":"array","items":{"$ref":"#/components/schemas/IngredientReq"}},"cat":{"$ref":"#/components/schemas/Category"},"orig":{"$ref":"#/components/schemas/Origin"}}},"UserType":{"title":"UserType","enum":["admin","user","guest"],"type":"string","description":"An enumeration."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}